+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 - check book loaned status before loaning
 - make tests more thorough
-- when user authenticated, add token to template
+- create view to handle user authentication and to include token to template
 - fix apparent csrf error when posting through front-end
 
 ## 2020-03-28